createBet controller

+++ Backend/models/Bet.go
@@ -4,17 +4,17 @@ import "gorm.io/gorm"
 
 type Bet struct {
 	gorm.Model
-	Winner *int
-	KoTko *int
-	Soum *int
-	R1 *int
-	R2 *int
-	R3 *int
-	R4 *int
-	R5 *int
-	Points *int
-	Draw *int
+	Winner  *int
+	KoTko   *int
+	Soum    *int
+	R1      *int
+	R2      *int
+	R3      *int
+	R4      *int
+	R5      *int
+	Points  *int
+	Draw    *int
 	MatchID string
-	GroupID uint
-	UserID uint
+	GroupID uint64
+	UserID  uint64
 }
